+++ Cesium3DTilesSelection/src/GltfModifierVersionExtension.cpp
@@ -3,7 +3,7 @@
 
 namespace Cesium3DTilesSelection {
 
-/* static */ std::optional<int32_t> GltfModifierVersionExtension::getVersion(
+/* static */ std::optional<int64_t> GltfModifierVersionExtension::getVersion(
     const CesiumGltf::Model& model) noexcept {
   const auto* pExtension = model.getExtension<GltfModifierVersionExtension>();
   return pExtension ? std::make_optional(pExtension->version) : std::nullopt;
@@ -11,7 +11,7 @@ namespace Cesium3DTilesSelection {
 
 /* static */ void GltfModifierVersionExtension::setVersion(
     CesiumGltf::Model& model,
-    int32_t version) noexcept {
+    int64_t version) noexcept {
   auto* pExtension = model.getExtension<GltfModifierVersionExtension>();
   if (!pExtension) {
     pExtension = &model.addExtension<GltfModifierVersionExtension>();
